requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
torch==2.4.1
torchvision==0.19.1
pillow==10.4.0
